test_requirements.txt
pytest>=2.9.0
docker-py
backoff
docker-compose==1.5.2
